+++ indicators/core/nlp_utils.py
@@ -9,10 +9,10 @@ from corextopic import vis_topic as vt
 from corextopic import corextopic as ct
 
 import pickle
-import pandas as pd
 from nesta.packages.nlp_utils.ngrammer import Ngrammer
 from sklearn.feature_extraction.text import CountVectorizer
 from indicators.core.config import MYSQLDB_PATH
+from indicators.core.core_utils import object_getter
 
 
 def join_text(*args):
@@ -160,7 +160,7 @@ def parse_corex_topics(path, n_most=5):
     return topics
 
 
-def fit_topic_model(model_config, object_getter):
+def fit_topic_model(topic_module, model_config):
     """Fit topics based on hyperparameters specified in the model config.
 
     Args:
@@ -171,7 +171,7 @@ def fit_topic_model(model_config, object_getter):
         objects, topic_model: List of objects (articles or projects),
                               and a trained topic model
     """
-    objs = next(object_getter())  # only one value (a list), so use next
+    objs = next(object_getter(topic_module))  # only one value (a list), so use next
     texts = [obj["text"] for obj in objs]
     titles = [obj["title"] for obj in objs]
     # Prepare the data and fit the model
