+++ indicators/core/nuts_utils.py
@@ -90,18 +90,18 @@ def make_reverse_lookup(data, key=itemgetter(1), prefix=""):
 
 
 @lru_cache()
-def get_geo_lookup(lat_lon_getter, iso2_to_id_getter):
+def get_geo_lookup(module):
     # Forward lookup
     nf = NutsFinder()
     id_to_nuts_lookup = {
-        id: nf.find(lat=lat, lon=lon) for id, lat, lon in lat_lon_getter()
+        id: nf.find(lat=lat, lon=lon) for id, lat, lon in module.lat_lon_getter()
     }
     id_nuts = [  # splatten out the nuts IDs, ready for grouping
         (id, info["NUTS_ID"])
         for id, nuts_info in id_to_nuts_lookup.items()
         for info in nuts_info
     ]
-    id_iso2 = iso2_to_id_getter()
+    id_iso2 = module.iso2_to_id_getter()
     # Reverse lookups
     nuts_to_id_lookup = make_reverse_lookup(id_nuts)
     iso2_to_id_lookup = make_reverse_lookup(id_iso2, prefix="iso_")
